requirements.txt
moviepy==1.0.3
google-generativeai==0.8.3
requests==2.31.0
python-dotenv==1.0.1
Pillow==10.4.0